fixed pointer issues now to fix physics update

+++ project_code/Tentacles_Testing_2/src/VerletPhysics/includes/VerletWorld.h
@@ -32,8 +32,8 @@ class VerletWorld
     VerletWorld();
     ~VerletWorld();
     
-    double previousTime;
-    double currentTime;
+    long previousTime;
+    long currentTime;
     int fixedDeltaTime;
     float fixedDeltaTimeSeconds;
     int leftOverDeltaTime;
